Deduce UriKind more intelligently in the Uri(string) constructor
If the supplied string starts with a backslash or a dot, presumably this
is intended to be a relative URI. This allows `new Uri("/foo")` to work
instead of throwing an exception about the URI being of an unrecognized
format.

Also assumes a URI to be relative if it begins with a literal dot, such
as `./Foo` or `../Bar`. A schema-less URI such as `Baz` is not treated
as a relative URI however, as it's not possible to definitively
deterimne if that is a schema-less path to a hostname without a TLD
or if it is intended to refer to a relative URI with that page name.
While the odds are the latter is the case (and that's how `<a
href="foo">...</a>` would be treated by a browser), `UriKind.Absolute`
is still used in this case to preserve backwards compatibility.